Add phone number hashing

+++ src/main/java/repository/UserRepository.java
@@ -3,6 +3,7 @@ package repository;
 import model.User;
 import model.UserGender;
 import model.UserType;
+import org.apache.commons.codec.digest.DigestUtils;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -50,15 +51,34 @@ public class UserRepository {
 
     public boolean save(User user) throws SQLException {
         Connection connection = getConnection();
+
+        int phoneNumberToInt = Integer.parseInt(user.getPhoneNumber());
+        String fakePhoneNumber = Integer.toString((phoneNumberToInt + 102030405));
+        user.setPhoneNumber(fakePhoneNumber);
+
+        String hashFromRealNumber = DigestUtils.sha256Hex(user.getPhoneNumber());
+
+        PreparedStatement fakePhoneNumberStatement = connection.prepareStatement(
+                "INSERT INTO `competence-schema`.`fake_phones` " +
+                        "(fake_phone_number, real_phone_number) VALUES (?, ?)");
+        fakePhoneNumberStatement.setString(1, fakePhoneNumber);
+        fakePhoneNumberStatement.setString(2, hashFromRealNumber);
+        boolean isFinished =  fakePhoneNumberStatement.executeUpdate() > 0;
+
+        if (!isFinished) {
+            return false;
+        }
+
         PreparedStatement statement = connection.prepareStatement(
                 "INSERT INTO `competence-schema`.`persons` (id, phone_number, profile_name, experiment_id, user_gender, user_age) VALUES (?, ?, ?, ?, ?, ?)");
         statement.setString(1, user.getUserID().toString());
-        statement.setString(2, user.getPhoneNumber());
+        statement.setString(2, fakePhoneNumber);
         statement.setString(3, user.getUserType().toString());
         statement.setString(4, user.getExperimentId());
         statement.setString(5, user.getUserGender().name());
         statement.setInt(6, user.getUserAge());
-        boolean isFinished =  statement.executeUpdate() > 0;
+        isFinished =  statement.executeUpdate() > 0;
+
         connection.close();
         return isFinished;
     }
@@ -81,10 +101,23 @@ public class UserRepository {
 
     public boolean delete(UUID id) throws SQLException {
         Connection connection = getConnection();
+
+        String fakeNumber = getById(id).getPhoneNumber();
+
+        PreparedStatement phoneStatement = connection.prepareStatement("DELETE FROM `competence-schema`.`fake_phones` WHERE fake_phone_number=?");
+        phoneStatement.setString(1, fakeNumber);
+
+        boolean isFinished =  phoneStatement.executeUpdate() > 0;
+
+        if (!isFinished) {
+            return false;
+        }
+
         PreparedStatement statement = connection.prepareStatement("DELETE FROM `competence-schema`.`persons` WHERE id=?");
         statement.setString(1, id.toString());
 
-        boolean isFinished =  statement.executeUpdate() > 0;
+        isFinished =  statement.executeUpdate() > 0;
+
         connection.close();
         return isFinished;
     }
